agents/agent2/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0
google-generativeai>=0.3.1
pydantic>=2.0.0
python-dotenv>=1.0.0
